move LiveServer class from fixtures.py to a separate module
